Cancellation and Checking in Mechanics Added

+++ src/main/java/com/campusEvent/campus_event/entity/Event.java
@@ -28,7 +28,7 @@ public class Event {
 
     @Column private LocalTime startTime, endTime;
 
-    @Version private int version;
+    @Version private Long version;
     @Column
     @Enumerated(EnumType.STRING)
     private EventStatus eventStatus;
